+++ src/pages/BecomeMember/BecomeMember.jsx
@@ -1,8 +1,7 @@
 import React from "react";
 import AppointmentPage from "./AppointmentSlide";
 import FormAppointment from "@/components/ui/FormAppointment";
-import NewsAppointment from "./NewsAppointment";
-import AfterNewsMembership from "./lastPageAppointment";
+import CombinedPage from "./CombinedPage";
 import { AboutAppointment } from "./AboutAppointment";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 const queryClient = new QueryClient();
@@ -15,8 +14,7 @@ export const BecomeMember = () => {
       <QueryClientProvider client={queryClient}>
         <FormAppointment />
       </QueryClientProvider>
-      <NewsAppointment />
-      <AfterNewsMembership />
+      <CombinedPage />
     </div>
   );
 };
